Add new Target Quality Metrics (#988)
The goal is to expand Target Quality
mode with multiple metrics besides VMAF. As an introduction, I added
SSIMULACRA2 and Butteraugli. ~~I added XPSNR as an option as well but it
is yet to be implemented.~~ (Done) A few other changes were also made to
support future work on this feature - primarily the refactor to the new
"metrics" crate.~~ A small bug was fixed for `--probing-rate` - this can
be reverted or separated out if necessary.~~ (Handled by different PR)

The metric can be chosen via a new parameter: `--target-metric [vmaf,
ssimulacra2, butteraugli-inf, butteraugli-3, xpsnr, xpsnr-weighted]`.
Here is an example command:

```ps
> av1an -i input.mkv -w 1 -e svt-av1 --sc-downscale-height 720 -y --target-quality 70 --target-metric ssimulacra2 --probing-rate 4
```

This will perform Target Quality using SSIMULACRA2 as the metric rather
than VMAF on probes 1/4 of the length.

## Operation

The majority of the feature requires VapourSynth and up to 3 additional
VapourSynth plugins listed below. The `-V --version` output has been
updated to reflect these new dependencies.

* [Vapoursynth-HIP (vship)](https://github.com/Line-fr/Vship) - Used for
HW accelerated SSIMULACRA2 and Butteraugli
* [VapourSynth Zig Image
Process](https://github.com/dnjulek/vapoursynth-zip) - CPU-based
SSIMULACRA2 and XPSNR
*
[vapoursynth-julek-plugin](https://github.com/dnjulek/vapoursynth-julek-plugin)
- CPU-based Butteraugli

The latter CPU-based implementations are used as a fallback when
VapourSynth-HIP is unavailable. With the exception of XPSNR, they are
not recommended but merely provided for compatibility. All of these
VapourSynth methods are slower than VMAF via FFmpeg but their utility
makes Target Quality mode a viable option for more users.

## Caveats and Issues

There are many and perhaps equally as many unaccounted for.

* ~~New external dependencies required - documentation needs to be
updated~~
* ~~VapourSynth chunking methods are required in addition to FFmpeg~~
(Documented)
* ~~`Select` chunking method currently scores incorrectly - needs
further investigation~~ (FFmpeg produces incompatible probe outputs that
cannot be used in VapourSynth)
* ~~Running multiple workers could result in vapoursynth-rs locking the
`loadscript.vpy` and crashing intermittently (AKA non-deterministic
behavior)~~ (Resolved by using `script_text` workaround)
* SSIMULACRA2 and Butteraugli scores are averaged instead of using
percentile due to the volatility of scores frame-to-frame
* ~~Butteraugli scores are seemingly inaccurate during probing - needs
further investigation~~ (Resolved by PR #1002)
* ~~vapoursynth-rs is now outdated - consider its replacement which
requires rewriting all these new functions~~ (Unnecessary)

The majority of changes were made in `vapoursynth.rs` and I'm not
confident that these new functions are appropriately handled. Exceptions
that have occurred when calling plugin functions incorrectly crash in
vapoursynth-rs/mod.rs and provide no meaningful feedback. While this is
an annoyance during development, it's even worse when an enduser cannot
provide useful logs. For example, the function that imports a video via
BestSource assumes a recent version of BestSource is installed, allowing
for setting the "cachemode" parameter to something unsupported in older
versions. There is no way to accurately know what version of BestSource
is installed so this function may need to be modified to somehow handle
that exception and try again without the "cachemode" parameter.

In other words, I don't know the best practices for handling exceptions
in functions that return `anyhow`.

## Next steps

### Address caveats and issues

The surface area of Target Quality has been increased dramatically with
the addition of VapourSynth as a dependency for these new metrics.
Exception handling, documentation, and command validation are all
necessary to make the application more robust. It may be necessary to
introduce a dedicated function to validate that the commandline
arguments and ensure there is no chance of an error occurring.

The file locking issue with vapoursynth-rs is not exclusive to Target
Quality mode. I have had this issue occur when running multiple workers
with a very complex or long-running VapourSynth script. I don't fully
understand why or how it happens either; just that a worker/chunk fails
because it can't find `loadscript.vpy` and occasionally crashes the
encoder. However, a lot of these new changes to `vapoursynth.rs` also
provide us the opportunity to finally remove `loadscript.vpy` (and
perhaps Python as a dependency) altogether and using vapoursynth-rs to
create and evaluate VapourSynth scripts in runtime and directly pipe the
output to the encoder. That may alleviate the issue but it's better to
address that separately.

Update: Scripts are now evaluated in runtime and no longer relies on
constant filesystem access to `loadscript.vpy`.

### More metrics

PSNR, SSIM, and XPSNR should be implemented just like how VMAF is
already implemented via FFmpeg. Other metrics are available in other
VapourSynth plugins but there may be a lack of demand to justify the dev
time.

The new "metrics" crate is a much-needed refactor to better organize the
codebase and introduce more metrics to plot on SVGs. This could be
separated out into a different PR but the VMAF module already shares
some functions between plotting and Target Quality. More metrics added
could do the same or be further refactored in the pursuit of DRY.

### Advanced Probing

Update: These suggestions have been migrated to their respective
Issues/PRs.

Probing is currently done with 2 key parameters: `--probing-rate` and
`--probe-slow`. However, there are other strategies that can be
implemented in order to speed up probing or improve accuracy.

Probing rate could be expanded to the maximum defined by the minimum
scene length rather than just 4. `--probe-slow` should be deprecated for
2 new parameters: `--probe-speed [veryslow, slow, medium, fast,
veryfast]` and `--probe-custom` which appends the custom settings for
the probe which can then be modified to trade speed for accuracy.
`--probe-slow` can be supported by silently mapping it to
`--probe-custom` as it performs the same task. Adjusting the
speed/accuracy separately addresses some Issues reported where an
encoder like SVT produces blocking on the fastest preset (which is what
is currently used). Users can currently work around that bug by using
`--probe-slow` but that may shift the scale of the issue the other way.
With `--probe-speed`, users can provide the `fast` or lower to avoid the
bug and maintain most of the performance without coupling the encoding
speed to the probing speed.

Another strategy is to probe subsections of the chunk rather than the
modulus of N frames. For example, rather than probing 100 of 400 frames
in a chunk, probe the first or middle or last N frames of the chunk.

So far the methods described were temporal. Resizing the width/height
could allow us to probe spatially. However, this may not be viable as
quantizer values may not scale in such a way.

Thanks,
\- Boats M.

---------

Co-authored-by: Boats McGee <BoatsMcGee@users.noreply.github.com>
Co-authored-by: Josh Holmer <jholmer.in@gmail.com>
Co-authored-by: Emre AKYÜZ <emreakyuz_2@hotmail.com>