Fix CLI argument validation warnings do not print (#1022)
Resolves #1003

Instead of bailing (exit with error message), present the user with a
warning message (see below) that suggests they install mkvmerge or
specify a different concatenation method. This allows users to perform
Scene Detection without having to also specify a different concatenation
method, which is unrelated to the user's intent.

The `tracing::warn` import was missing from the `settings.rs` file, so
this PR also fixes that.

```ps
> av1an.exe -i input.mkv --temp temporary --concat mkvmerge --sc-only
 WARN mkvmerge not found, but `--concat mkvmerge` was specified. Make sure to install mkvmerge or specify a different concatenation method (e.g. `--concat ffmpeg`) before encoding.
 INFO resume was set but neither chunks.json nor done.json exist in temporary directory temporary
 INFO encode_file: Input: 1920x1080 @ 23.976 fps, YUV420P10LE, SDR
Scene detection
00:00:06 ▐████████████████████████████████████████████████████████████████████████████▌ 100% 452/452 (73.83 fps, eta 0s) INFO encode_file: scenecut: found 5 scene(s) [with extra_splits (239 frames): 5 scene(s)]
```

There are many other checks that can fail similarly (such as a missing
default "aomenc" encoder), but this PR does not handle every case where
this could occur. For that much effort, it would be better spent on
rewriting the CLI.

Ideally, this would fall under the large refactor mentioned in Feature
Request #995 where the CLI is reorganized. For this specific
application, a subcommand would be a better pattern for handling Scene
Detection and nothing else. That command would have its own validation
that does not rely on extraneous checks.

---------

Co-authored-by: Boats McGee <BoatsMcGee@users.noreply.github.com>