Fix final probe copy shortcut ignores probing rate, probing speed, and ffmpeg filters (#1023)
Resolves #1021

During Target Quality, if the user specifies a `--probing-speed` with
`--probe-slow`, the final probe is copied for the final encode chunk
despite possibly being a different speed/preset than the one specified.
Similarly, when the user specifies an FFmpeg filter parameter, the
shortcut does not account for it as Target Quality probing does not use
FFmpeg filters. This PR adds a check to ensure both of these values are
unspecified.

Update: This PR now adds a check for the probing rate to ensure a value
none other than `1` is used.

---------

Co-authored-by: Boats McGee <BoatsMcGee@users.noreply.github.com>