Add Target Quality Retries (#1026)
Resolves #993
Addresses #787 but does not resolve it - not all suggested features
accepted or implemented.

This PR adds retry logic to Target Quality in the same way as encoding.
If Target Quality fails `--max-retries` (default 3) times in a single
chunk, it will log the error and exit. This should alleviate any
transient issues with FFmpeg or an encoder when probing.

Thanks,
\- Boats M.

---------

Co-authored-by: Boats McGee <BoatsMcGee@users.noreply.github.com>