Further improvements for CRF search
With this change, the spline crate is also removed. Some simple functions are added for linear, natural cubic and PCHIP interpolations.

Subject to be improved further in the future (if we come up with even better methods). The current method will be:

Start with binary search (we do it already): First Round (with 0 known data point)
Binary again for the second round (we do it already): Second round (with 1 known data points)
Linear interpolation (we do it already): Third round (with 2 known data points)
Natural Cubic Spline: Fourth round (with 3 known data points)
For the 5th probe, use PCHIP Interpolation (with 4 known data points)
At this point (6 and rest rounds) only do binary search until the exit condition (because generally this point indicates a problem and it should finish faster with this). This means we exhausted all the scientific answers and it didn't work.