Fix MKVMerge Fails to Concatenate on Linux When Total Chunks Exceed ulimit (#1065)
Resolves #547
Resolves #760

## Overview

Users on Linux are limited by the amount of file descriptors that can be
opened by a single `mkvmerge` process (or the user itself). A workaround
of setting the `-ulimit -n` value to one that well exceeds the total
number of chunks to merge and performing the merge manually is not a
long-term solution and may not work for all environments. Instead, the
chunks can be merged into groups of 100 chunks that are then merged into
the final output. This group size should be well within reasonable
limits.

Additionally, the `String::with_capacity(VALUE)` has been adjusted to
account for varying input lengths. I merely printed the capacity (in
bytes) before and the length (returns bytes) afterwards to ensure there
was no reallocation occurring when writing to the `file_string`.

### Testers Wanted

I have not tested the failure case on Linux, only Windows. Windows
either has no such limit or I have yet to reach it in my testing. If you
have encountered this issue or are willing to test with an environment
where the `ulimit -n` is set very low, please give this PR a try to make
absolutely sure this avoids the `ulimit` issue.

### Potential Improvements

Group sizes could be determined by checking the `ulimit -n` value rather
than hardcoding to 100.

Currently, the groups are executed linearly but could be parallelized
with threads assuming `ulimit` only applies to a single process rather
than the entire shell or user environment.

The groups are redundant and could be deleted, keeping the temporary
folder disk usage low.

The merging preprocess is ephemeral and video files can be very large.
To reduce disk thrashing, the groups should be stored in memory until
the final output is merged and written to the disk. Doing so across
platforms is a difficult task in and of itself and could be even more
difficult if no such solution exists for Rust.

Thank you,
\- Boats M.

---------

Co-authored-by: Boats McGee <BoatsMcGee@users.noreply.github.com>