Fix Split Method None Zoning Logic
Zones are sorted and have no overlap so the frames_processed value needs to be updated to the last frame processed rather than adding up the end frame of every zone. Additionally guard against the first check to add an empty scene when frames_processed is initialized at 0 and the first zone is being evaluated. This condition is to add a scene between zones - essentially frames that were not included in the specified zones. Checking against the end_frame is incorrect as we want to use the start of the zone instead to evaluate the condition.