Reduce Scene Detection loadscript Usage (#1030)
Resolves #1029
Addresses #994 but does not completely resolve it. This change only
applies to Scene Detection and does not remove the original
`loadscript.vpy`, only the copied Scene Detection loadscript.

~~Requires rust-av/av-scenechange#182 and so the crate has been updated
temporarily to use the fork/branch for that PR for demonstration
purposes.~~ Merged.
Requires rust-av/av-decoders#1. Make sure to update
`av1an-core/cargo.toml` to the latest `av-scenechange` and `av-decoders`
before merging this PR.

Uses the underlying rust-av/av-decoders#1 changes to modify the output
video node for downscaling programmatically rather than blindly
modifying the user-provided input VapourSynth script. `Input` has been
modified to store the necessary details to generate a loadscript text
each time it is needed. For user-provided VapourSynth scripts, the path
is used to instantiate
`av-decoders::VapoursynthDecoder::from_file(Path)`, which will ensure
the CWD for the script is the script directory itself. This does not
apply for Target Quality probe measurement because using
`environment.eval_file()` causes occasional File System issues due to
frequently accessing the VapourSynth script file. During validation, if
there is a discrepancy between the input script location and the CWD, we
present a new warning recommending the script be in the same location as
the CWD.

For non-VapourSynth inputs that use a VapourSynth chunking method such
as FFMS2, the script can be generated and passed to
`av-decoders::VapourSynthDecoder::from_script()` just like the
user-provided script. Generating at the endpoint reduces the log
pollution recently introduced in PR #988 when the `script_text` property
was added. Note: The user-provided script is not generated/read from
disk every time it is requested. Doing so would thrash the file system
with frequent reads so we instead cache the value when initializing
`Input::VapourSynth`. The aforementioned "log pollution" is not resolved
for this particular input for this reason.

`register_node_modifier` takes a callback that lets us modify the output
node programmatically and return it to use instead of the node evaluated
from the just the script. This allows us to programmatically apply
downscaling without resorting to copying and modifying a user-provided
script with the risk of an invalid modification.

Current Scene Detection Decoding Scenarios:

1. User-provided VapourSynth script inputs with `--sc-downscale-height`
specified: Apply downscaling using the most reliable method,
`VapoursynthDecoder::from_file(Path)` and
`register_node_modifier(callback)`
2. User-provided VapourSynth script inputs with no downscaling or Video
inputs with a VapourSynth chunking method: Use VSPipe with provided
script or loadscript as before
3. Y4M video remains the same: using FFmpeg to downscale when `filters`
vector is not empty

### Additional Changes

`clip_info` caching requires us to retrieve the most accurate info the
first time, so the input using a VapourSynth chunk method will be cached
during validation when `clip_info` is first called. For some inputs,
this could take a long time. I added an `info!()` log when this happens
but was not able to see it. When DGDecNV is cached, we already had a log
so the user has some feedback but the aforementioned `info!()` log does
not show, despite initializing the logging system before validation.
LSMASH and FFMS2 are fast enough that it may not be an issue, but
BestSource indexing/caching is a lengthy process with no feedback.

The loadscript.vpy has been updated to ensure the downscale height is an
int. The core thread limit has been removed for Scene Detection only -
the limit still applies during encoding/probing. With a core thread
limit Scene Detection decode speeds are negatively impacted.

### Caveats

Due to the slower decoding speed from `VapoursynthDecoder`, it is only
used when we need it for the reliability. The following details only
apply when the user provides their own VapourSynth script as input along
with a Scene Detection downscale height.

Currently, `av-decoders` appears not to take advantage of the entire
systems resources, leading to a slower than current Scene Detection
performance. This could potentially be resolved in `av-decoders` and/or
`av-scenechange`. However, the system resource usage with this PR appear
to be significantly lower. Here is some testing and data below:

The following were using **`-m ffms2 --sc-downscale-height 720`** on
`1920x1080 @ 23.976 fps, YUV420P10LE, SDR`.
Before (`vspipe.exe loadscript.vpy | av1an y4m`):
`00:00:30 ▌  10% 3253/34095 (107.77 fps, eta 5m)`
After (`av-scenechange` + `av-decoders::VapoursynthDecoder` +
`register_node_modifier()`):
`00:00:30 ▌   5% 1856/34095 (60.87 fps, eta 9m)`

**`--sc-downscale-height 540`**
Before:
`00:00:30 ▌  12% 3993/34095 (131.61 fps, eta 4m)`
After:
`00:00:30 ▌   8% 2596/34095 (85.67 fps, eta 6m)`

I ran a similarly long test and monitored my resource usage in Task
Manager as a quick sanity check on general performance.
Before: Total usage is ~80% of my entire machine, which is all of the
available resources, presumably limited by CPU. Av1an uses ~10% CPU,
vspipe uses 70%.
After: Total usage is ~40% of my entire machine, which is *not* all of
the available resources, presumably limited by implementation. Av1an
uses anywhere from 35%-50% but mostly around 40%.

Given how low this resource usage is, I believe Av1an could be
significantly faster if we resolve the decoding speed in the underlying
dependency. Av1an code should need no change for when that happens
besides bumping the dependency version.

Thanks,
\- Boats M.

---------

Co-authored-by: Boats McGee <BoatsMcGee@users.noreply.github.com>
Co-authored-by: Josh Holmer <jholmer.in@gmail.com>