Add more pedantic clippy lints
The most significant change is disallowing `.unwrap()`. Instead, we
should:

- If the error is one that we never expect to occur in production (i.e.
  it indicates a bug), we should use `.expect` to add reasoning why we
  believe this won't occur.
- If the error is expected to be triggerable in production, we should
  bubble it up with `?` instead of using `.unwrap()`. Unwrap in
  production is bad for a variety of reasons, so we should avoid it.

There are also a number of lints to improve code readability, use more
idiomatic patterns, and improve performance. The reason some of these
lints are opt-in is because they are restrictions that may not apply to
all code bases. There are a few where we are only going to enable them
for the core library and not for the CLI, for example. Some others are
still experimental and may have false positives. Instances of false
positives can be ignored using `#[expect(clippy::lint_name, reason =
"...")]`.