fix: remove git lfs
Although git lfs would be ideal for storing the y4m file used in the
integration tests, it is a pain in the rear. In addition to the
additional cloning steps required, it is apparently breaking the AUR
package for av1an-git. Therefore, removing it seems to be the simplest
way forward.