Fix FFmpeg VMAF fails to validate even when Target Quality is not used (#1103)
Resolves #1102

Target Metric VMAF is now only validated if the user supplies
`--target-quality` or specifies at least 1 Zone where Target Quality is
enabled. Zones are now validated after parsing so any issues with a
specified Target Metric can be addressed in the new function.

I initially wanted to cache the validation as I planned to validate each
Zone as it's parsed but instead went with parsing all Zones, then doing
a filter check for the exact condition(s) that require validation. At
worst, XPSNR could be validated up to 3 times, 2 for everything else.
With the function only executing up to a maximum of 3 times, I decided
to forgo complicating things with lazy loading and caching.

Thanks,
\- Boats M.

---------

Co-authored-by: Boats McGee <BoatsMcGee@users.noreply.github.com>